+++ packages/guide/src/app/npcs/npcs.module.ts
@@ -18,6 +18,7 @@ import { MatInput } from "@angular/material/input";
 import { MatOption } from "@angular/material/autocomplete";
 import { MatSelect } from "@angular/material/select";
 import { MatCheckbox } from "@angular/material/checkbox";
+import { NpcFilterComponent } from "./npc-filter/npc-filter.component";
 
 @NgModule({
     declarations: [
@@ -29,7 +30,7 @@ import { MatCheckbox } from "@angular/material/checkbox";
         HeartEventTriggerComponent,
         GiftingGridComponent,
     ],
-    imports: [CommonModule, NPCsRoutingModule, SharedModule, MatTooltipModule, MatProgressSpinnerModule, AddSpacesToPascalCasePipe, MatFormField, ReactiveFormsModule, MatInput, MatLabel, MatOption, MatSelect, MatCheckbox],
+    imports: [CommonModule, NPCsRoutingModule, SharedModule, MatTooltipModule, MatProgressSpinnerModule, AddSpacesToPascalCasePipe, MatFormField, ReactiveFormsModule, MatInput, MatLabel, MatOption, MatSelect, MatCheckbox, NpcFilterComponent],
 })
 export class NPCsModule {
 }